Move check_beamstop into commmon
