+++ src/mx_bluesky/common/device_setup_plans/check_beamstop.py
@@ -1,5 +1,5 @@
 import bluesky.plan_stubs as bps
-from dodal.devices.i03.beamstop import Beamstop, BeamstopPositions
+from dodal.devices.mx_phase1.beamstop import Beamstop, BeamstopPositions
 
 from mx_bluesky.common.utils.log import LOGGER
