+++ src/mx_bluesky/hyperion/device_setup_plans/utils.py
@@ -8,14 +8,14 @@ from dodal.devices.detector import (
 )
 from dodal.devices.detector.detector_motion import DetectorMotion, ShutterState
 from dodal.devices.eiger import EigerDetector
-from dodal.devices.i03.beamstop import Beamstop
+from dodal.devices.i03 import Beamstop
 from dodal.devices.i03.dcm import DCM
 
+from mx_bluesky.common.device_setup_plans.check_beamstop import check_beamstop
 from mx_bluesky.common.device_setup_plans.position_detector import (
     set_detector_z_position,
     set_shutter,
 )
-from mx_bluesky.hyperion.device_setup_plans.check_beamstop import check_beamstop
 
 
 def fill_in_energy_if_not_supplied(dcm: DCM, detector_params: DetectorParams):
