+++ tests/system_tests/hyperion/external_interaction/test_load_centre_collect_full_plan.py
@@ -17,6 +17,7 @@ from ophyd_async.core import AsyncStatus
 from ophyd_async.testing import set_mock_value
 from PIL import Image
 
+from mx_bluesky.common.device_setup_plans.check_beamstop import BeamstopException
 from mx_bluesky.common.external_interaction.callbacks.common.ispyb_mapping import (
     get_proposal_and_session_from_visit_string,
 )
@@ -30,7 +31,6 @@ from mx_bluesky.common.utils.exceptions import (
     CrystalNotFoundException,
     WarningException,
 )
-from mx_bluesky.hyperion.device_setup_plans.check_beamstop import BeamstopException
 from mx_bluesky.hyperion.experiment_plans.load_centre_collect_full_plan import (
     LoadCentreCollectComposite,
     load_centre_collect_full,
